backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # Not needed for SQLite

# AI Integration
google-generativeai==0.3.2

# HTTP & Validation
httpx==0.26.0
python-multipart==0.0.6
email-validator==2.1.0

# Environment & Config
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# Background Tasks & Scheduling
apscheduler==3.10.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
